Fix text wraparound draw issue on ILI9341
This commit partially fixes the text wraparound issue w/ the ILI9341 -- this is accomplished by using the screen dimensions along with the font size to calculate the maximum possible characters that can be printed.

If the max char length is exceeded, the rest of the string is truncated.

An issue still exists on this commit with screen clipping, where the right-hand portion of the display does not draw text, this is due to a draw/screen window that is set incorrectly somewhere. This will need to be adjusted, for the time being a hardcoded value of 50 has been applied to adjust for this region.